Fixes on Encoder and Test FrontEnd
Some issues fixed.
Remaining:
Will the encrypt method ever return a newline or null character?
Note: To access the Test FrontEnd, add a backslash to the beginning of
the first line, and delete the first backslash on line 54. If we decide
to just use this test FrontEnd, I labeled a section on the test FrontEnd
where the layout of everything can be changed.